+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format of this changelog is based on
 
 ## [Unreleased]
 
+- (feat) Add support for pnpm-workspace.yaml in pnpm-setup.
 - Update workflows to support nodejs 24
 
 ## [0.37.0] <a name="0.37.0" href="#0.37.0">-</a> January 14, 2025
@@ -600,7 +601,6 @@ latest on the `master` branch.
 - Deprecate `call_main`: use `run_main`
 - `run_main` allows us to handle the results and issues.
 - `m ci lint` supports the output of
-
   - eslint
   - pycodestyle
   - pylint